+++ src/main/java/org/cqfn/astranaut/core/base/DefaultPosition.java
@@ -1,0 +1,72 @@
+/*
+ * The MIT License (MIT)
+ *
+ * Copyright (c) 2024 Ivan Kniazkov
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included
+ * in all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NON-INFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+package org.cqfn.astranaut.core.base;
+
+/**
+ * Default implementation of the Position interface that fits in most cases.
+ * @since 2.0.0
+ */
+public final class DefaultPosition implements Position {
+    /**
+     * Object representing the source code that has this position.
+     */
+    private final Source source;
+
+    /**
+     * The row (line) number of this position.
+     */
+    private final int row;
+
+    /**
+     * The column number of this position.
+     */
+    private final int column;
+
+    /**
+     * Constructor.
+     * @param source Object representing the source code that has this position
+     * @param row The row (line) number of this position
+     * @param column The column number of this position
+     */
+    public DefaultPosition(final Source source, final int row, final int column) {
+        this.source = source;
+        this.row = row;
+        this.column = column;
+    }
+
+    @Override
+    public Source getSource() {
+        return this.source;
+    }
+
+    @Override
+    public int getRow() {
+        return this.row;
+    }
+
+    @Override
+    public int getColumn() {
+        return this.column;
+    }
+}
